app/requirements.txt
fastapi
uvicorn[standard]

# Templates & static
jinja2
aiofiles

# Database
sqlalchemy
asyncpg

# Auth
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart
email-validator

# Config
python-dotenv

# Redis
redis

# Observability
prometheus-client
python-json-logger

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-sqlalchemy